Time: 4 ms (4.90%) | Memory: 23.9 MB (22.85%) - LeetSync

+++ 637-average-of-levels-in-binary-tree/average-of-levels-in-binary-tree.cpp
@@ -28,7 +28,7 @@ if(node->left) q.push(node->left);
 if(node->right) q.push(node->right);
 
       }
-      ans.push_back((double)sum/(float)n);}
+      ans.push_back((double)sum/n);}
       return ans;
     }
 };
